+++ src_bonus/collectibe_heart/collectible_heart_set_bonus.c
@@ -48,7 +48,7 @@ void	set_heart_cords(t_game *game)
 	while (y < game->height)
 	{
 		x = 0;
-		while (x < ft_strlen(game->map[y]))
+		while (x < (int)ft_strlen(game->map[y]))
 		{
 			if (game->map[y][x] == 'H')
 			{
