+++ src_bonus/collectible_ammo/collectible_ammo_set_bonus.c
@@ -50,7 +50,7 @@ void	set_c_ammo_cords(t_game *game)
 	while (y < game->height)
 	{
 		x = 0;
-		while (x < ft_strlen(game->map[y]))
+		while (x < (int)ft_strlen(game->map[y]))
 		{
 			if (game->map[y][x] == 'A')
 			{
