+++ src_bonus/map_checks/flood_fill_bonus.c
@@ -50,7 +50,7 @@ static void	process_flood_fill_b(t_game *game, int **visited, \
 		{
 			nx = x + dxy[i];
 			ny = y + dxy[i + 1];
-			if (ny < game->height && nx < ft_strlen(game->map[ny]))
+			if (ny < game->height && nx < (int)ft_strlen(game->map[ny]))
 			{
 				if (is_valid_b(game, visited, nx, ny))
 					stack_op_b(stack, &nx, &ny, PUSH);
